fix(rpc): Switch measurements to float (#83192)
- They always should've been floats, but at the initial implementation
we didn't have float arrays yet, now that we do updating it

+++ src/sentry/search/eap/constants.py
@@ -61,8 +61,7 @@ TYPE_MAP: dict[SearchType, AttributeKey.Type.ValueType] = {
     "week": FLOAT,
     "duration": FLOAT,
     "integer": INT,
-    # TODO:  need to update these to float once the proto supports float arrays
-    "number": INT,
+    "number": FLOAT,
     "percentage": FLOAT,
     "string": STRING,
 }
